split styleguide test files
