+++ tests/playwright/sanity/modules/styleguide/styleguide-1.test.ts
@@ -1,11 +1,8 @@
-import { expect, Page, TestInfo } from '@playwright/test';
+import { expect, Page } from '@playwright/test';
 import { parallelTest as test } from '../../../parallelTest';
 import WpAdminPage from '../../../pages/wp-admin-page';
-import ApiRequests from '../../../assets/api-requests';
 
 test.describe( 'Styleguide Preview tests @styleguide_image_link', () => {
-	const fontsContentText = 'The five boxing wizards jump quickly.';
-
 	test.beforeAll( async ( { browser, apiRequests }, testInfo ) => {
 		const page = await browser.newPage();
 		const wpAdmin = new WpAdminPage( page, testInfo, apiRequests );
@@ -144,6 +141,32 @@ test.describe( 'Styleguide Preview tests @styleguide_image_link', () => {
 		await expect( styleguidePreviewDialog ).toBeHidden();
 	} );
 
+	test( 'Switching between tabs makes relevant area visible', async ( { page, apiRequests }, testInfo ) => {
+		// Arrange.
+		const { editor } = await getInSettingsTab( page, testInfo, apiRequests, 'Global Colors', true );
+		await page.waitForTimeout( 2000 );
+
+		await page.locator( '#elementor-panel-header-kit-back' ).click();
+		await styleguideSaveChanges( page );
+		await page.waitForTimeout( 2000 );
+
+		// Act.
+		await page.click( '.elementor-panel-menu-item-title:has-text("Global Fonts")' );
+
+		// Assert
+		await expect( editor.getPreviewFrame().getByText( 'Global Fonts' ) ).toBeVisible();
+
+		// Act 2.
+		await page.locator( '#elementor-panel-header-kit-back' ).click();
+		await styleguideSaveChanges( page );
+		await page.waitForTimeout( 2000 );
+
+		await page.click( '.elementor-panel-menu-item-title:has-text("Global Colors")' );
+
+		// Assert 2
+		await expect( editor.getPreviewFrame().getByText( 'Global Colors' ) ).toBeVisible();
+	} );
+
 	test( 'Enabling Styleguide Preview at Global Colors shows the Styleguide Modal and updates user preferences', async ( { page, apiRequests }, testInfo ) => {
 		// Arrange.
 		const { editor } = await getInSettingsTab( page, testInfo, apiRequests, 'Global Colors', false );
@@ -271,252 +294,24 @@ test.describe( 'Styleguide Preview tests @styleguide_image_link', () => {
 		// Assert 3.
 		await expect( primaryColor ).not.toHaveClass( /active/ );
 	} );
-
-	test( 'Clicks on font trigger picker state and active state', async ( { page, apiRequests }, testInfo ) => {
-		// Arrange.
-		const { editor } = await getInSettingsTab( page, testInfo, apiRequests, 'Global Fonts', true );
-
-		const secondaryFont = editor.getPreviewFrame().getByText( 'Secondary' + fontsContentText );
-		const picker = page.locator( '.elementor-control-popover-toggle-toggle-label' ).nth( 1 );
-
-		// Act 1. Click on font in preview.
-		await secondaryFont.click();
-
-		// Assert 1.
-		await expect( secondaryFont ).toHaveClass( /active/ );
-		expect( await page.getByText( 'Typography' ).nth( 1 ).isVisible() ).toBeTruthy();
-
-		// Act 2. click on font again nothing happens.
-		await secondaryFont.click();
-
-		// Assert 2.
-		await expect( secondaryFont ).toHaveClass( /active/ );
-		expect( await page.getByText( 'Typography' ).nth( 1 ).isVisible() ).toBeTruthy();
-
-		// Arrange 3.
-		await secondaryFont.click();
-
-		// Act 3. Click picker after already active.
-		await picker.click();
-
-		// Assert 3.
-		await expect( secondaryFont ).not.toHaveClass( /active/ );
-	} );
-
-	test( 'Change font title', async ( { page, apiRequests }, testInfo ) => {
-		// Arrange.
-		const { editor } = await getInSettingsTab( page, testInfo, apiRequests, 'Global Fonts', true );
-
-		const input = page.locator( '.elementor-repeater-fields' ).nth( 2 );
-
-		// Act - Click on font in preview.
-		await input.click();
-		await input.type( 'more' );
-
-		// Assert
-		const TextFont = editor.getPreviewFrame().getByText( 'Textmore' + fontsContentText );
-		await expect( TextFont ).toContainText( 'Textmore' );
-	} );
-
-	test( 'Change color title', async ( { page, apiRequests }, testInfo ) => {
-		// Arrange.
-		const { editor } = await getInSettingsTab( page, testInfo, apiRequests, 'Global Colors', true );
-
-		const input = page.locator( '.elementor-repeater-fields' ).nth( 1 );
-
-		// Act - Change title.
-		await input.click();
-		await input.type( 'more' );
-
-		// Assert
-		const secondaryColor = editor.getPreviewFrame().getByText( /Secondarymore#[A-Fa-f0-9]{6}/i );
-		await expect( secondaryColor ).toContainText( 'Secondarymore' );
-	} );
-
-	test( 'Adding and removing new colors', async ( { page, apiRequests }, testInfo ) => {
-		// Arrange.
-		const { editor } = await getInSettingsTab( page, testInfo, apiRequests, 'Global Colors', true );
-
-		const picker = page.locator( '.elementor-repeater-fields' ).nth( 4 ).locator( '.pcr-button' );
-		const addButton = page.getByRole( 'button', { name: 'Add Color' } );
-
-		// Act.
-		await addButton.click();
-		await picker.click();
-		await page.locator( '.pcr-result' ).nth( 4 ).fill( '#594833' );
-		await picker.click();
-
-		// Assert
-		expect( await editor.getPreviewFrame().getByText( /New Item #1#594833/i ).count() ).toEqual( 1 );
-
-		// Arrange 2.
-		const listItem = page.locator( '.elementor-repeater-fields' ).nth( 4 ).getByText( 'Reorder Remove' );
-		const remove = listItem.locator( '.eicon-trash-o' );
-
-		// Act 2 - Click on remove.
-		await listItem.hover();
-		await remove.click();
-		await page.getByRole( 'button', { name: 'Delete' } ).click();
-
-		// Assert 2
-		const number = await editor.getPreviewFrame().getByText( /New Item #1#/i ).count();
-		expect( number ).toEqual( 0 );
-	} );
-
-	test( 'Adding and removing new fonts', async ( { page, apiRequests }, testInfo ) => {
-		// Arrange.
-		const { editor } = await getInSettingsTab( page, testInfo, apiRequests, 'Global Fonts', true );
-		const addButton = page.getByRole( 'button', { name: 'Add Style' } );
-
-		// Act.
-		await addButton.click();
-
-		// Assert
-		const customFontsCount = await page.locator( '.elementor-control-custom_typography .elementor-repeater-fields' ).count();
-		expect( await editor.getPreviewFrame().getByText( 'New Item #' + customFontsCount + fontsContentText ).count() ).toEqual( 1 );
-
-		// Arrange 2.
-		const listItem = page.locator( '.elementor-control-custom_typography .elementor-repeater-fields' ).last().getByText( 'Edit Remove Reorder' );
-		const remove = listItem.locator( '.eicon-trash-o' );
-
-		// Act 2 - Click on remove.
-		await listItem.hover();
-		await remove.click();
-		await page.getByRole( 'button', { name: 'Delete' } ).click();
-
-		// Assert 2
-		const number = await editor.getPreviewFrame().getByText( 'New Item #' + customFontsCount + fontsContentText ).count();
-		expect( number ).toEqual( 0 );
-	} );
-
-	test( 'Changed color in picker to reflect in styleguide', async ( { page, apiRequests }, testInfo ) => {
-		// Arrange.
-		const { editor } = await getInSettingsTab( page, testInfo, apiRequests, 'Global Colors', true );
-		const secondaryColor = editor.getPreviewFrame().getByText( /Secondary#[A-Fa-f0-9]{6}/i );
-
-		const picker = page.locator( '.elementor-repeater-fields' ).nth( 1 ).locator( '.pcr-button' );
-		// Act.
-		await picker.click();
-		await page.locator( '.pcr-result' ).nth( 1 ).fill( '#594833' );
-
-		// Assert
-		await expect( secondaryColor.locator( 'div' ).first() ).toHaveCSS( 'background-color', 'rgb(89, 72, 51)' );
-	} );
-
-	test( 'Changed font values in picker to reflect in styleguide', async ( { page, apiRequests }, testInfo ) => {
-		// Arrange.
-		const { editor } = await getInSettingsTab( page, testInfo, apiRequests, 'Global Fonts', true );
-
-		const textFont = editor.getPreviewFrame().getByText( 'Text' + fontsContentText );
-		const picker = page.locator( '.elementor-repeater-fields' ).nth( 2 ).locator( '.eicon-edit' ).first();
-
-		// Act.
-		await picker.click();
-		const select = page.locator( '.elementor-repeater-fields' ).nth( 2 ).getByRole( 'combobox', { name: 'Style' } );
-		await select.selectOption( { label: 'Italic' } );
-		const select2 = page.locator( '.elementor-repeater-fields' ).nth( 2 ).getByRole( 'combobox', { name: 'Decoration' } );
-		await select2.selectOption( { label: 'Underline' } );
-
-		// Assert
-		const content = textFont.locator( 'p' ).nth( 1 );
-		await expect( content ).toHaveCSS( 'font-style', 'italic' );
-		await expect( content ).toHaveCSS( 'text-decoration-line', 'underline' );
-	} );
-
-	test( 'Switching between tabs makes relevant area visible', async ( { page, apiRequests }, testInfo ) => {
-		// Arrange.
-		const { editor } = await getInSettingsTab( page, testInfo, apiRequests, 'Global Colors', true );
-		await page.waitForTimeout( 2000 );
-
-		await page.locator( '#elementor-panel-header-kit-back' ).click();
-		await styleguideSaveChanges( page );
-		await page.waitForTimeout( 2000 );
-
-		// Act.
-		await page.click( '.elementor-panel-menu-item-title:has-text("Global Fonts")' );
-
-		// Assert
-		await expect( editor.getPreviewFrame().getByText( 'Global Fonts' ) ).toBeVisible();
-
-		// Act 2.
-		await page.locator( '#elementor-panel-header-kit-back' ).click();
-		await styleguideSaveChanges( page );
-		await page.waitForTimeout( 2000 );
-
-		await page.click( '.elementor-panel-menu-item-title:has-text("Global Colors")' );
-
-		// Assert 2
-		await expect( editor.getPreviewFrame().getByText( 'Global Colors' ) ).toBeVisible();
-	} );
-
-	test( 'Clicking header buttons makes relevant area visible', async ( { page, apiRequests }, testInfo ) => {
-		// Arrange.
-		const { editor } = await getInSettingsTab( page, testInfo, apiRequests, 'Global Colors', true );
-
-		const fontsButton = editor.getPreviewFrame().getByRole( 'button', { name: 'Fonts' } );
-		const colorButton = editor.getPreviewFrame().getByRole( 'button', { name: 'Colors' } );
-
-		// Act.
-		await fontsButton.click();
-
-		// Assert
-		await expect( editor.getPreviewFrame().getByText( 'Global Fonts' ) ).toBeVisible();
-
-		// Act 2.
-		await colorButton.click();
-
-		// Assert 2
-		await expect( editor.getPreviewFrame().getByText( 'Global Colors' ) ).toBeVisible();
-	} );
 } );
 
 async function isStyleguidePreviewUserPreferencesEnabled( page: Page ) {
 	return await page.evaluate( () => elementor.getPreferences( 'enable_styleguide_preview' ) );
 }
 
-async function getInSettingsTab( page: Page, testInfo: TestInfo, apiRequests: ApiRequests, tabName: string, styleguideOpen: boolean ) {
-	const wpAdmin = new WpAdminPage( page, testInfo, apiRequests );
-	const editor = await wpAdmin.openNewPage();
-	page.setDefaultTimeout( 10000 );
-
-	await page.evaluate( ( isOpen ) => $e.run( 'document/elements/settings', {
-		container: elementor.settings.editorPreferences.getEditedView().getContainer(),
-		settings: {
-			enable_styleguide_preview: isOpen ? 'yes' : '',
-		},
-		options: {
-			external: true,
-		},
-	} ), styleguideOpen );
-
-	await page.waitForTimeout( 3000 );
-
-	await Promise.all( [
-		page.waitForResponse( '/wp-admin/admin-ajax.php' ),
-		editor.openSiteSettings( ),
-	] );
-
-	await page.waitForTimeout( 1000 );
-
-	await page.click( `.elementor-panel-menu-item-title:has-text("${ tabName }")` );
-
-	await page.waitForTimeout( 1000 );
-
-	return { editor, wpAdmin };
-}
-
-async function styleguideSaveChanges( page ) {
+async function styleguideSaveChanges( page: Page ) {
 	for ( const dialog of await page.locator( '.dialog-confirm-widget-content' ).all() ) {
-		if ( ! await dialog.isVisible( 1000 ) ) {
+		if ( ! await dialog.isVisible() ) {
 			continue;
 		}
 
-		const dialogHeader = await dialog.locator( '.dialog-header' );
-		if ( ! dialogHeader.count() || ! ( await dialogHeader.innerText() ).includes( 'Save Changes' ) ) {
+		const dialogHeader = dialog.locator( '.dialog-header' );
+		if ( ! await dialogHeader.count() || ! ( await dialogHeader.innerText() ).includes( 'Save Changes' ) ) {
 			continue;
 		}
 
-		const saveButton = await dialog.locator( '.dialog-button:has-text("Save")' );
+		const saveButton = dialog.locator( '.dialog-button:has-text("Save")' );
 		await saveButton.click();
 	}
 }
